refactored ``MockResponse``
