+++ pytest_response/interceptors/urllib_quick.py
@@ -3,6 +3,7 @@ import urllib.request
 from functools import wraps
 
 from pytest_response import response
+from pytest_response.app import BaseMockResponse
 from pytest_response.exceptions import RemoteBlockedError, ResponseNotFound
 from pytest_response.logger import log
 
@@ -32,44 +33,9 @@ def urlopen_wrapper(func):
     return inner_func
 
 
-class MockResponse:
+class MockResponse(BaseMockResponse):
     def __init__(self, data, headers={}):
-        self.status = self.code = 200
-        self.msg = self.reason = "OK"
-        self.headers = headers
-        self.will_close = True
-        if not isinstance(data, io.BytesIO):
-            data = io.BytesIO(data)
-        self.fp = data
-        self.will_close = True
-
-    def flush(self):
-        self.fp.flush()
-
-    def info(self):
-        return {}
-
-    def read(self, *args, **kwargs):
-        """
-        Wrapper for _io.BytesIO.read
-        """
-        return self.fp.read(*args, **kwargs)
-
-    def readline(self, *args, **kwargs):
-        """
-        Wrapper for _io.BytesIO.readline
-        """
-        return self.fp.readline(*args, **kwargs)
-
-    def readinto(self, *args, **kwargs):
-        """
-        Wrapper for _io.BytesIO.readinto
-        """
-        return self.fp.readinto(*args, **kwargs)
-
-    def close(self):
-        if hasattr(self, "fp"):
-            self.fp.close()
+        super().__init__(data, headers)
 
     pass
 
@@ -83,7 +49,6 @@ def install_opener():
 
 def uninstall_opener():
     response.mpatch.undo()
-    # response.mpatch.setattr("urllib.request.urlopen", )
 
 
 install = install_opener
